Remove framework from supported versions (see description)
We were having issues with System.Text.Json not existing in framework and the `#if !NETFRAMEWORK` seemingly not working anymore in newer versions of dotnet. The easiest option for us it to just remove framework for the list of support for now. We don't use framework for any of our apps that depend on this fork.